Add ReTest to other packages
