+++ BanyanArrays/Project.toml
@@ -17,6 +17,7 @@ Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 AWSS3 = "1c724243-ef5b-51ab-93f4-b0a88ac62a95"
+ReTest = "e0db7c4e-2690-44b9-bad6-7687da720f89"
 
 [targets]
-test = ["Banyan", "Test", "Dates", "Distributions", "HDF5", "AWSS3"]
+test = ["Banyan", "Test", "ReTest", "Dates", "Distributions", "HDF5", "AWSS3"]
