+++ BanyanDataFrames/Project.toml
@@ -29,6 +29,7 @@ Parquet = "626c502c-15b0-58ad-a749-f091afb673ae"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+ReTest = "e0db7c4e-2690-44b9-bad6-7687da720f89"
 
 [targets]
-test = ["Arrow", "AWSS3", "Banyan", "BanyanArrays", "CSV", "DataFrames", "FilePathsBase", "Parquet", "Statistics", "Test", "Random"]
+test = ["Arrow", "AWSS3", "Banyan", "BanyanArrays", "CSV", "DataFrames", "FilePathsBase", "Parquet", "Statistics", "Test", "ReTest", "Random"]
